fix fb share submission and redirect

+++ web/modules/custom/glico_submission/src/Controller/AjaxSubmissionController.php
@@ -4,6 +4,7 @@ namespace Drupal\glico_submission\Controller;
 
 use Drupal\Core\Ajax\AjaxResponse;
 use Drupal\Core\Ajax\CloseModalDialogCommand;
+use Drupal\Core\Ajax\RedirectCommand;
 use Drupal\Core\Controller\ControllerBase;
 use Drupal\node\Entity\Node;
 use Symfony\Component\HttpFoundation\Response;
@@ -20,33 +21,22 @@ class AjaxSubmissionController extends ControllerBase {
    *   Return Hello string.
    */
   public function submission($flg) {
+    $response = new AjaxResponse();
+    $tempstore = \Drupal::service('user.private_tempstore')
+      ->get('glico_submission');
+    $nid = $tempstore->get('nid');
+    $node = Node::load($nid);
+    if ($node === NULL) {
+      return $response;
+    }
+
     if ($flg == 'not-shared') {
-      $tempstore = \Drupal::service('user.private_tempstore')
-        ->get('glico_submission');
-      $file = $tempstore->get('file');
-      $frame = $tempstore->get('frame');
-      $baby_name = $tempstore->get('baby_name');
-      $caption = $tempstore->get('caption');
-      $nodes = \Drupal::entityTypeManager()
-        ->getStorage('node')
-        ->loadByProperties([
-          'type' => 'submission',
-          'title' => $baby_name,
-          'field_frame' => $frame,
-          'field_caption' => $caption,
-          'field_video' => [
-            'target_id' => $file,
-          ],
-        ]);
-      $node = array_shift($nodes);
       $node->delete();
     }
 
-    $response = new AjaxResponse();
-    $response->addCommand(new CloseModalDialogCommand());
+    if ($flg == 'shared') {
+      $response->addCommand(new RedirectCommand($node->toUrl()->toString()));
+    }
     return $response;
   }
-
-
-
 }
